+++ web_src/src/ui/configurationFieldRenderer/ObjectFieldRenderer.tsx
@@ -244,6 +244,20 @@ export const ObjectFieldRenderer: React.FC<FieldRendererProps> = ({
     [schemaDefaults, value],
   );
 
+  // When value is missing or empty object, push schema defaults to parent so required-object
+  // validation (e.g. "schedule is required") sees a non-empty value and doesn't flag the field.
+  const hasPushedSchemaDefaults = React.useRef(false);
+  React.useEffect(() => {
+    if (hasPushedSchemaDefaults.current) return;
+    const isEmpty =
+      value === undefined ||
+      value === null ||
+      (typeof value === "object" && value !== null && Object.keys(value).length === 0);
+    if (!isEmpty || Object.keys(schemaDefaults).length === 0) return;
+    hasPushedSchemaDefaults.current = true;
+    onChange(schemaDefaults);
+  }, [value, schemaDefaults, onChange]);
+
   return (
     <div className="border border-gray-300 dark:border-gray-700 rounded-md p-4 space-y-4">
       {schema.map((schemaField) => (
